Add files via upload
This is the restaurant menu. It is not fully automatic, but rather semi-automatic. It is an electronic menu, which means you can access it on your device. Once you have decided what to order, you can call the waiter by clicking on a button provided on the menu. We can customize the menu by editing the prices, food items, and descriptions according to the restaurant's preferences. The menu is user-friendly, with easy navigation and a visually appealing design.

https://www.linkedin.com/posts/rashmi-kiran-uidesignercumapoet_codsoft-uxdesign-uidesign-activity-7119009905872711680-APuW?utm_source=share&utm_medium=member_desktop