Add files via upload
This is a small e-commerce website that sells teddy bears. The website offers three sizes of teddy bears: small, medium, and large. Customers can easily place an order and customize the color and other attributes based on the website owner's available options. There is a user profile page, along with sections for About Us and Policy. Overall, it is a simple and convenient platform for buying teddy bears.

https://www.linkedin.com/posts/rashmi-kiran-uidesignercumapoet_codsoft-uxdesign-uidesign-activity-7119010965928472576-yMOh?utm_source=share&utm_medium=member_desktop